[fix] Don't crash on missing timestamps for releases
It's possible that npm is missing times for packages as seen in
nodejitsu/browsenpm.or#25. The suspected cause for this is that these
releases are pre-releases or tagged releases and they don't get
processed by npm in the same way as regular releases. As some of the
data is missing.